+++ src/LanguageServer/Impl/Indexing/IndexParser.cs
@@ -13,34 +13,30 @@
 // See the Apache Version 2.0 License for specific language governing
 // permissions and limitations under the License.
 
-using System;
-using System.Diagnostics;
 using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Python.Core.Diagnostics;
 using Microsoft.Python.Core.IO;
 using Microsoft.Python.Parsing;
+using Microsoft.Python.Parsing.Ast;
 
 namespace Microsoft.Python.LanguageServer.Indexing {
     internal sealed class IndexParser : IIndexParser {
-        private readonly ISymbolIndex _symbolIndex;
         private readonly IFileSystem _fileSystem;
         private readonly PythonLanguageVersion _version;
         private readonly CancellationTokenSource _allProcessingCts = new CancellationTokenSource();
         private readonly object _syncObj = new object();
         private CancellationTokenSource _linkedParseCts;
 
-        public IndexParser(ISymbolIndex symbolIndex, IFileSystem fileSystem, PythonLanguageVersion version) {
-            Check.ArgumentNotNull(nameof(symbolIndex), symbolIndex);
+        public IndexParser(IFileSystem fileSystem, PythonLanguageVersion version) {
             Check.ArgumentNotNull(nameof(fileSystem), fileSystem);
 
-            _symbolIndex = symbolIndex;
             _fileSystem = fileSystem;
             _version = version;
         }
 
-        public Task ParseAsync(string path, CancellationToken cancellationToken = default) {
+        public Task<PythonAst> ParseAsync(string path, CancellationToken cancellationToken = default) {
             lock (_syncObj) {
                 CancelCurrentParse();
                 _linkedParseCts = CancellationTokenSource.CreateLinkedTokenSource(_allProcessingCts.Token, cancellationToken);
@@ -56,27 +52,21 @@ namespace Microsoft.Python.LanguageServer.Indexing {
             _linkedParseCts = null;
         }
 
-        private void Parse(string path, CancellationTokenSource parseCts) {
+        private PythonAst Parse(string path, CancellationTokenSource parseCts) {
             parseCts.Token.ThrowIfCancellationRequested();
-            try {
-                using (var stream = _fileSystem.FileOpen(path, FileMode.Open, FileAccess.Read, FileShare.Read)) {
-                    var parser = Parser.CreateParser(stream, _version);
-                    var ast = parser.ParseFile();
-                    lock (_syncObj) {
-                        parseCts.Token.ThrowIfCancellationRequested();
-                        _symbolIndex.Add(path, ast);
-                    }
-                }
-            } catch (Exception e) when (e is IOException || e is UnauthorizedAccessException) {
-                Trace.TraceError(e.Message);
-            } finally {
-                lock (_syncObj) {
-                    if (_linkedParseCts == parseCts) {
-                        _linkedParseCts.Dispose();
-                        _linkedParseCts = null;
-                    }
+            PythonAst ast = null;
+            using (var stream = _fileSystem.FileOpen(path, FileMode.Open, FileAccess.Read, FileShare.Read)) {
+                var parser = Parser.CreateParser(stream, _version);
+                ast = parser.ParseFile();
+            }
+            parseCts.Token.ThrowIfCancellationRequested();
+            lock (_syncObj) {
+                if (_linkedParseCts == parseCts) {
+                    _linkedParseCts.Dispose();
+                    _linkedParseCts = null;
                 }
             }
+            return ast;
         }
 
         public void Dispose() {
